make flit and packet ids as long long

+++ simulator/src/model/traffic/Flit.cpp
@@ -23,9 +23,9 @@
 #include "Packet.h"
 #include "TrafficPool.h"
 
-int Flit::idcnt = 0;
+long long Flit::idcnt = 0;
 
-Flit::Flit(FlitType type, int seq_nb, Packet* p)
+Flit::Flit(FlitType type, long long seq_nb, Packet* p)
         :
         type(type),
         seq_nb(seq_nb),
@@ -45,7 +45,7 @@ Flit::Flit(FlitType type, int seq_nb, Packet* p)
         this->headFlit = p->flits.at(0);
 }
 
-Flit::Flit(FlitType type, int seq_nb, Packet* p, dataTypeID_t dataType, double generationTime)
+Flit::Flit(FlitType type, long long seq_nb, Packet* p, dataTypeID_t dataType, double generationTime)
         :
         Flit(type, seq_nb, p)
 {
